+++ zellij-server/src/tab/layout_applier.rs
@@ -180,6 +180,11 @@ impl<'a> LayoutApplier<'a> {
                     }
                 }
                 pane_focuser.focus_tiled_pane(&mut self.tiled_panes);
+                LayoutApplier::offset_viewport(
+                    self.viewport.clone(),
+                    self.tiled_panes,
+                    self.draw_pane_frames,
+                );
             },
             Err(e) => {
                 Err::<(), _>(anyError::msg(e))
@@ -200,6 +205,7 @@ impl<'a> LayoutApplier<'a> {
         let free_space = self.total_space_for_tiled_panes();
         match layout.position_panes_in_space(&free_space, None) {
             Ok(positions_in_layout) => {
+                let mut run_instructions_to_ignore = layout.run_instructions_to_ignore.clone();
                 let positions_and_size = positions_in_layout.iter();
                 let mut new_terminal_ids = new_terminal_ids.iter();
 
@@ -211,8 +217,17 @@ impl<'a> LayoutApplier<'a> {
                 };
 
                 for (layout, position_and_size) in positions_and_size {
-                    // A plugin pane
-                    if let Some(Run::Plugin(run)) = layout.run.clone() {
+                    if let Some(position) = run_instructions_to_ignore
+                        .iter()
+                        .position(|r| r == &layout.run)
+                    {
+                        let run = run_instructions_to_ignore.remove(position);
+                        self.tiled_panes.set_geom_for_pane_with_run(
+                            run,
+                            *position_and_size,
+                            layout.borderless,
+                        );
+                    } else if let Some(Run::Plugin(run)) = layout.run.clone() {
                         let pane_title = run.location.to_string();
                         let pid = new_plugin_ids
                             .get_mut(&run.location)
@@ -320,10 +335,15 @@ impl<'a> LayoutApplier<'a> {
         let mut new_floating_terminal_ids = new_floating_terminal_ids.iter();
         for floating_pane_layout in floating_panes_layout {
             layout_has_floating_panes = true;
-            if let Some(Run::Plugin(run)) = floating_pane_layout.run.clone() {
-                let position_and_size = self
-                    .floating_panes
-                    .position_floating_pane_layout(&floating_pane_layout);
+            let position_and_size = self
+                .floating_panes
+                .position_floating_pane_layout(&floating_pane_layout);
+            if floating_pane_layout.already_running {
+                self.floating_panes.set_geom_for_pane_with_run(
+                    floating_pane_layout.run.clone(),
+                    position_and_size,
+                );
+            } else if let Some(Run::Plugin(run)) = floating_pane_layout.run.clone() {
                 let pane_title = run.location.to_string();
                 let pid = new_plugin_ids
                     .get_mut(&run.location)
@@ -363,9 +383,6 @@ impl<'a> LayoutApplier<'a> {
                     focused_floating_pane = Some(PaneId::Plugin(pid));
                 }
             } else if let Some((pid, hold_for_command)) = new_floating_terminal_ids.next() {
-                let position_and_size = self
-                    .floating_panes
-                    .position_floating_pane_layout(&floating_pane_layout);
                 let next_terminal_position =
                     get_next_terminal_position(&self.tiled_panes, &self.floating_panes);
                 let initial_title = match &floating_pane_layout.run {
@@ -504,28 +521,43 @@ impl<'a> LayoutApplier<'a> {
         self.tiled_panes.resize(new_screen_size);
         Ok(())
     }
-    fn offset_viewport(&mut self, position_and_size: &Viewport) {
-        let mut viewport = self.viewport.borrow_mut();
-        if position_and_size.x == viewport.x
-            && position_and_size.x + position_and_size.cols == viewport.x + viewport.cols
+    pub fn offset_viewport(
+        viewport: Rc<RefCell<Viewport>>,
+        tiled_panes: &mut TiledPanes,
+        draw_pane_frames: bool,
+    ) {
+        let boundary_geoms = tiled_panes.non_selectable_pane_geoms_inside_viewport();
         {
-            if position_and_size.y == viewport.y {
-                viewport.y += position_and_size.rows;
-                viewport.rows -= position_and_size.rows;
-            } else if position_and_size.y + position_and_size.rows == viewport.y + viewport.rows {
-                viewport.rows -= position_and_size.rows;
+            // curly braces here is so that we free viewport immediately when we're done
+            let mut viewport = viewport.borrow_mut();
+            for position_and_size in boundary_geoms {
+                if position_and_size.x == viewport.x
+                    && position_and_size.x + position_and_size.cols == viewport.x + viewport.cols
+                {
+                    if position_and_size.y == viewport.y {
+                        viewport.y += position_and_size.rows;
+                        viewport.rows -= position_and_size.rows;
+                    } else if position_and_size.y + position_and_size.rows
+                        == viewport.y + viewport.rows
+                    {
+                        viewport.rows -= position_and_size.rows;
+                    }
+                }
+                if position_and_size.y == viewport.y
+                    && position_and_size.y + position_and_size.rows == viewport.y + viewport.rows
+                {
+                    if position_and_size.x == viewport.x {
+                        viewport.x += position_and_size.cols;
+                        viewport.cols -= position_and_size.cols;
+                    } else if position_and_size.x + position_and_size.cols
+                        == viewport.x + viewport.cols
+                    {
+                        viewport.cols -= position_and_size.cols;
+                    }
+                }
             }
         }
-        if position_and_size.y == viewport.y
-            && position_and_size.y + position_and_size.rows == viewport.y + viewport.rows
-        {
-            if position_and_size.x == viewport.x {
-                viewport.x += position_and_size.cols;
-                viewport.cols -= position_and_size.cols;
-            } else if position_and_size.x + position_and_size.cols == viewport.x + viewport.cols {
-                viewport.cols -= position_and_size.cols;
-            }
-        }
+        tiled_panes.set_pane_frames(draw_pane_frames);
     }
     fn adjust_viewport(&mut self) -> Result<()> {
         // here we offset the viewport after applying a tiled panes layout
@@ -541,11 +573,11 @@ impl<'a> LayoutApplier<'a> {
             *display_area
         };
         self.resize_whole_tab(display_area).context(err_context)?;
-        let boundary_geoms = self.tiled_panes.borderless_pane_geoms();
-        for geom in boundary_geoms {
-            self.offset_viewport(&geom)
-        }
-        self.tiled_panes.set_pane_frames(self.draw_pane_frames);
+        LayoutApplier::offset_viewport(
+            self.viewport.clone(),
+            self.tiled_panes,
+            self.draw_pane_frames,
+        );
         Ok(())
     }
     fn set_focused_tiled_pane(&mut self, focus_pane_id: Option<PaneId>, client_id: ClientId) {