+++ zellij-utils/src/kdl/mod.rs
@@ -916,6 +916,9 @@ impl TryFrom<(&KdlNode, &Options)> for Action {
             },
             "PreviousSwapLayout" => Ok(Action::PreviousSwapLayout),
             "NextSwapLayout" => Ok(Action::NextSwapLayout),
+            "BreakPane" => Ok(Action::BreakPane),
+            "BreakPaneRight" => Ok(Action::BreakPaneRight),
+            "BreakPaneLeft" => Ok(Action::BreakPaneLeft),
             _ => Err(ConfigError::new_kdl_error(
                 format!("Unsupported action: {}", action_name).into(),
                 kdl_action.span().offset(),